Moved field logic to adapter
