+++ src/senaite/core/datamanagers/__init__.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
-#
+
 # flake8:noqa
-#
 # Backwards compatibility
 from .content import *
 from .field import *
