+++ src/senaite/core/datamanagers/content/__init__.py
@@ -1,1 +1,6 @@
 # -*- coding: utf-8 -*-
+
+# flake8:noqa
+from .analysis import RoutineAnalysisDataManager
+from .analysisservice import AnalysisServiceDataManager
+from .sample import SampleDataManager
