+++ src/senaite/core/datamanagers/field/configure.zcml
@@ -1,0 +1,10 @@
+<configure xmlns="http://namespaces.zope.org/zope">
+
+  <!-- Data Manager for Sample Analyses Field -->
+  <adapter
+      for="bika.lims.interfaces.IAnalysisRequest
+           zope.publisher.interfaces.browser.IBrowserRequest
+           Products.Archetypes.interfaces.IField"
+      factory=".sample_analyses.SampleAnalysesFieldDataManager" />
+
+</configure>
